Add unencrypted uplink (#2)
* Add unencrypted uplink

Creates an initial version of unencrypted pouch with an uplink.

This generally follows the architecture outlined in discussions in
golioth/architecture#61, with the exception of starting out with a push
based API towards the application, instead of a callback based one.

Instead of a callback per path, we're providing an event handler
registration mechanism that lets the application listen for session
started/ended events, and write buffers to the uplink using the
`pouch_uplink_entry_write()` function. The intent behind this change is
to make it easier for the application to allocate and push CBOR data,
which generally can't operate on partial buffers. Splitting this
interaction into events and write calls additionally lets the
application schedule packets outside of connections, and aligns better
with the API in the SDK. The downside of this change is that it
introduces some ambiguity around when the application no longer has any
data to send.

This PR introduces all the basic modules required for unencrypted pouch
transport, as well as some stubs and architectual nubs to allow an
encrypted mode to be added in the future.

Note that the CBOR based headers are not generated from CDDL, but done
inline. The CBOR handling has been isolated so that we can change this
in the future, but with the low complexity here, it isn't worth doing
IMO.

Signed-off-by: Trond Snekvik <trond@golioth.io>

* Variable length blocks

Based on whiteboarding session, this patch changes the blocks in the
pouch to variable length, which lets us make a few significant changes
to the code's structure.

- All uplink data goes into dedicated buffers that are allocated on the
  heap, and passed around in linked list queues.
- The ring buffer has been removed in favor of linked list buffers.
- The pouch header as well as each block is allocated in a dedicated
  buffer. As soon as an entry is unable to fit in a block, the block
  gets queued for processing and a new block is allocated.
- Processing is performed asynchronously in a work handler.
- Processing is only performed while a pouch is open, but multiple
  blocks can be queued up for processing ahead of time.
- Pouch headers are generated from CDDL.
- Entry and block headers are binary encoded.
- We perform block allocations in a blocking manner to simplify future
  stream mechanisms, which would get real flaky for long running streams
  without it.

Signed-off-by: Trond Snekvik <trond@golioth.io>
Co-authored-by: Sam Friedman <132283695+sam-golioth@users.noreply.github.com>