Add stream blocks
Adds an API for streaming data over the uplink. The streams are
heap allocated, and the application is responsible for their life cycle.
The application can write data to the stream through a posix like write
API.

Blocks are allocated for the stream as we go, with a single entry header
in the first block. The block type has been extended to a 1 + 7 bit
field, where the MSb is a flag indicating no more data. and the other
bits for indicating a block ID, where block ID 0 is reserved for entry
blocks. The no more data flag is always set for the entry blocks.

The streams are only valid for the pouch they were created for, and the
application is responsible for finishing the open streams before they
close the pouch. Streams that aren't closed before closing a pouch
are considered invalid, and will be lost.

Signed-off-by: Trond Snekvik <trond@golioth.io>