Create pouch_ble_gatt_common library
This commit breaks out the common functionality of the BLE GATT
transport (the parts that are useful for both GATT servers and
clients) into a new library. This consists of the packetizer and
the UUID definitions.

This commit also adds build system machinery for creating
library code that can be included into applications without
including the main `pouch` functionality.

The primary goal of this commit is to make it easier to share code
between nodes, which use both `pouch` and its transports, and
gateways, which only use transports.

Signed-off-by: Sam Friedman <sam@golioth.io>