block: Simplify header logic
Clarify the format of the block header. Include the header in the
configurable block size to simplify calculations. The header is
sent on the wire, so including it in the configurable block size
is more intuitive to the user. Previously, part, but not all, of
the header was allocated as additional overhead.

Signed-off-by: Sam Friedman <sam@golioth.io>