Use updated info characteristic
Changes the format of the info characteristic to flags and server cert
serial number. This allows us to remove the init function for the
attributes, as we have to construct the info structure before each read,
in case any of the data changes. This simplifies the macro, setup and
linker sections.

Renames the peripheral_create to init to reflect the lack of a destroy
function, now that we don't have to allocate anything.

Signed-off-by: Trond Snekvik <trond@golioth.io>