MbedTLS: enable parsing of certificates with 256 bit signature
The 256 bit signature in the server certificate can't be parsed from the
certificate data unless signature hash validation is enabled for the
specific hash in the certificate, as mbedTLS doesn't compile it in if it
doesn't think we need it. As a result, we wouldn't be able to parse the
certificate unless we enabled ECDSA and the desired hash algorithms in
mbedTLS, even if signature validation was disabled.

Signed-off-by: Trond Snekvik <trond@golioth.io>