saead: session: canonicalize blocksize representation
The blocksize is represented as a hexidecimal string during key
agreement. There are (at least) two ways to represent each value
that contains at least one alphabetic character because these
characters can be either upper or lower case. We have determined
that the canonical representation of hexadecimal numbers in strings
will use upper case, so this commit updates the implementation to
match.

Signed-off-by: Sam Friedman <sam@golioth.io>