downlink: fix 'pouch_buf' memory leak
'pouch_buf' is allocated in pouch_downlink_start() and possibly
reallocated in pouch_downlink_push(). However, there was no final
buf_free() call, so with the next pouch_downlink_start() there was a
memory leak due to pointer override with new allocation.

Free 'pouch_buf' in pouch_downlink_finish() and assing it to NULL, so
next invocation of pouch_downlink_start() is not creating memory leak.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>